Fix up plateform hitting the screen borders too soon

+++ src/main/java/be/bendem/gametest/core/graphics/Translatable.java
@@ -5,6 +5,6 @@ package be.bendem.gametest.core.graphics;
  */
 public interface Translatable {
 
-    public void translate(int x, int y);
+    public void translate(double x, double y);
 
 }
