+++ src/main/java/be/bendem/gametest/core/graphics/shapes/Circle.java
@@ -32,7 +32,7 @@ public class Circle extends Ellipse2D.Double implements GraphicObject {
     }
 
     @Override
-    public void translate(int x, int y) {
+    public void translate(double x, double y) {
         setFrame(getX() + x, getY() + y, getWidth(), getHeight());
     }
